src/components/Breadcrumb/index.spec.tsx
// src/components/Breadcrumb/index.test.tsx

import { render, screen } from '@testing-library/react'
import Breadcrumb from './index'

describe('Breadcrumb', () => {
  const items = [
    { name: 'Home', href: '/' },
    { name: 'Category', href: '/category' },
    { name: 'Product' },
  ]

  it('renders correctly with items', () => {
    render(<Breadcrumb items={items} />)

    const breadcrumbItems = screen.getAllByTestId('breadcrumb-item')
    const breadcrumbSeparators = screen.getAllByTestId('breadcrumb-separator')

    expect(breadcrumbItems).toHaveLength(items.length)
    expect(breadcrumbSeparators).toHaveLength(items.length - 1)
    expect(breadcrumbSeparators[0].textContent).toBe('>')
  })

  it('renders custom separator', () => {
    const items = [
      { name: 'Home', href: '/' },
      { name: 'Category', href: '/category' },
      { name: 'Product' },
    ]
    const separator = '🐟'

    render(<Breadcrumb items={items} separator={separator} />)

    const breadcrumbSeparators = screen.getAllByTestId('breadcrumb-separator')

    expect(breadcrumbSeparators[0].textContent).toBe(separator)
  })
})
